Move HIP DeepCopy to it's own files
